+++ IndyDotNet/Wallet/Async.cs
@@ -18,7 +18,7 @@ namespace IndyDotNet.Wallet
 #if __IOS__
         [MonoPInvokeCallback(typeof(OpenWalletCompletedDelegate))]
 #endif
-        private static void OpenWalletCallbackMethod(int xcommand_handle, int err, IntPtr wallet_handle)
+        private static void OpenWalletCallbackMethod(int xcommand_handle, int err, int wallet_handle)
         {
             var taskCompletionSource = PendingCommands.Remove<WalletAsync>(xcommand_handle);
 
@@ -298,13 +298,13 @@ namespace IndyDotNet.Wallet
         /// <summary>
         /// Gets the SDK handle for the Wallet instance.
         /// </summary>
-        internal IntPtr Handle { get; }
+        internal int Handle { get; }
 
         /// <summary>
         /// Initializes a new Wallet instance with the specified handle.
         /// </summary>
         /// <param name="handle">The SDK handle for the wallet.</param>
-        private WalletAsync(IntPtr handle)
+        private WalletAsync(int handle)
         {
             Handle = handle;
             _requiresClose = true;
